# [exporter/chronicleexporter]: Optimize getRawField when field is "body"